@@ -1,4 +1,4 @@
-import { Component, ElementRef, HostListener, ViewChild } from '@angular/core';
+import { Component, ElementRef, HostListener, viewChild } from '@angular/core';
 
 @Component({
   selector: 'app-about',
@@ -6,8 +6,7 @@ import { Component, ElementRef, HostListener, ViewChild } from '@angular/core';
   templateUrl: './about.component.html',
 })
 export class AboutComponent {
-  // References the full-width white container
-  @ViewChild('whiteFullWidthContainer') whiteFullWidthContainer!: ElementRef;
+  whiteFullWidthContainer = viewChild.required<ElementRef>('whiteFullWidthContainer');
 
   // Adjusts the fullwidth white container height after the view has been initialized
   ngAfterViewChecked() {
@@ -22,15 +21,15 @@ export class AboutComponent {
 
   // Check if body height is less than viewport height and adjust the fullwidth white container height to fill the screen
   private setContainerHeight() {
-    this.whiteFullWidthContainer.nativeElement.style.height = 'auto';
+    this.whiteFullWidthContainer().nativeElement.style.height = 'auto';
     const bodyHeight = document.body.clientHeight;
     const viewportHeight = window.innerHeight;
     if (bodyHeight < viewportHeight) {
-      const offsetTop = this.whiteFullWidthContainer.nativeElement.offsetTop;
+      const offsetTop = this.whiteFullWidthContainer().nativeElement.offsetTop;
       const height = viewportHeight - offsetTop;
-      this.whiteFullWidthContainer.nativeElement.style.height = `${height}px`;
+      this.whiteFullWidthContainer().nativeElement.style.height = `${height}px`;
     } else {
-      this.whiteFullWidthContainer.nativeElement.style.height = 'auto';
+      this.whiteFullWidthContainer().nativeElement.style.height = 'auto';
     }
   }
 }
