@@ -1,4 +1,4 @@
-import { Component, ElementRef, EventEmitter, Input, Output, ViewChild } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, Output, viewChild } from '@angular/core';
 import { NgClass } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { NgxCurrencyDirective } from 'ngx-currency';
@@ -10,6 +10,7 @@ import { NgxCurrencyDirective } from 'ngx-currency';
   templateUrl: './currency-input.component.html',
 })
 export class CurrencyInputComponent {
+  // Inputs and outputs
   @Input() label: string | undefined;
   @Input() symbol: string | undefined;
   @Input() placeholder: string | undefined;
@@ -18,19 +19,18 @@ export class CurrencyInputComponent {
   @Input() error: boolean | undefined;
   @Output() errorChange = new EventEmitter<boolean>();
   @Output() enterPressed = new EventEmitter<void>();
-  inputLeftPadding: string = '0.5rem';
 
-  // References the symbol div and the input element
-  @ViewChild('symbolDiv') symbolDiv!: ElementRef;
-  @ViewChild('input') input!: ElementRef;
+  // Reference symbolDiv and input elements
+  symbolDiv = viewChild.required<ElementRef>('symbolDiv');
+  input = viewChild.required<ElementRef>('input');
 
   // Set inputLeftPadding every time the view is checked
   ngAfterViewChecked() {
-    const offsetWidthInPx = this.symbolDiv.nativeElement.offsetWidth;
+    const offsetWidthInPx = this.symbolDiv().nativeElement.offsetWidth;
     const rootFontSize = parseFloat(getComputedStyle(document.documentElement).fontSize);
     const offsetWidthInRem = offsetWidthInPx / rootFontSize;
-    const inputLeftPadding = this.symbolDiv.nativeElement.offsetWidth ? offsetWidthInRem + 0.1875 + 'rem' : '0.5rem';
-    this.input.nativeElement.style.paddingLeft = inputLeftPadding;
+    const inputLeftPadding = this.symbolDiv().nativeElement.offsetWidth ? offsetWidthInRem + 0.1875 + 'rem' : '0.5rem';
+    this.input().nativeElement.style.paddingLeft = inputLeftPadding;
   }
 
   // Event handler for when the input amount changes
